+++ pysmt/smtlib/script.py
@@ -139,8 +139,8 @@ class SmtLibScript(object):
         assert self.count_command_occurrences(smtcmd.CHECK_SAT) == 1
         _And = mgr.And if mgr else And
 
-        assertions = (cmd.args
-                      for cmd in self.filter_by_command_name([smtcmd.ASSERT]))
+        assertions = [cmd.args[0]
+                      for cmd in self.filter_by_command_name([smtcmd.ASSERT])]
         return _And(assertions)
 
     def get_last_formula(self, mgr=None):