+++ docs/CHANGES.rst
@@ -37,6 +37,8 @@ General:
 * Removed option _quantified_ from all shortcuts.
 
 
+* Added PicoSAT solver wrapper
+
 
 0.2.2 2015-02-07 -- BDDs
 ------------------------